document-forgery-verification-system new/document-forgery-verification-system/requirements.txt
flask
opencv-python
pytesseract
pillow
numpy
pdf2image
